requirements.txt
aiogram
aiosqlite
httpx
python-dotenv
yookassa
fastapi
uvicorn
pyjwt
passlib[bcrypt]
pydantic[email]